Set up CLI menus
Slim down the CLI options to just the generic components we need to bootstrap a
CAPv cluster on vSphere.

This pulls in much of the logic and workflow of the nks-hci-vsphere code base,
but distills it down to just genconfig, deploy and destroy.

This commit is just the framework for the genconfig command.