Rework cli (#20)
* Rename package imports

Renaming repo to cake, so need to go through and fix things up a bit.

* Go back to a single binary

Move things into a single cake binary, we should be able to keep everything down
to a single utility program with flags for now.  If we have to change/expand
later we can.

Also added platform build options to the Makefile while adjusting.